#18 Update producer and consumer events

+++ kaftter/src/main/java/kaftter/vo/User.java
@@ -18,5 +18,5 @@ public class User {
     @JsonProperty("screen_name")
     private String screenName;
     @JsonProperty("followers_count")
-    private Long followers;
+    private int followers;
 }
